+++ core/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 1.0.6 (in progress)
 
+### Added
+
++ [#1784](https://github.com/luyadev/luya/issues/1784) Added auto cover sensitive data for error api transfer.
+
 ## 1.0.5 (26. March 2018)
 
 ### Changed